lab-1/python-flask/4-logs/requirements.txt
# App dependencies
flask==2.3.2

# OTel dependencies
opentelemetry-api==1.18.0
opentelemetry-sdk==1.18.0
opentelemetry-exporter-otlp==1.18.0
opentelemetry-instrumentation-logging==0.39b0
